+++ react/react-addons.d.ts
