+++ react/react-global.d.ts
